Included AVX512 commands to calculate squared distance faster Rewritten main to build MST for a set of points set in a text file

+++ point.h
@@ -1,7 +1,7 @@
 #include <vector>
 
-typedef std::vector<double> Point;
+typedef std::vector<float> Point;
 
-double Distance(const Point& a, const Point& b);
+float Distance(const Point& a, const Point& b);
 
 //Hello world
